Remove PrimeNumbers class, import as Maven dependency
The PrimeNumbers class was taken out of this project and used to create
a stand alone jar file with the mvn install command. This created a jar file
and saved it to the local Maven repository which is located at
~/.m2/repository.

The library was then added as dependency into the pom.xml file
and all local references were updated to refer to imported package.